experiments/token_2048_best_thinking/requirements.txt
# Token 2048 最佳思考实验依赖

# 核心依赖（verl已包含）
pandas>=2.0.0
numpy<2.0.0
datasets>=2.14.0
hydra-core>=1.3.0
ray[default]>=2.41.0
omegaconf>=2.3.0
pyyaml>=6.0  # 用于配置检查

# verl框架（从根目录安装）
# cd /mnt/nas/chenhaotian/verl && pip install -e .

# 可选依赖
matplotlib>=3.5.0  # 用于可视化
tqdm>=4.65.0  # 进度条

